+++ app/Http/Controllers/sanphamController.php
@@ -66,7 +66,9 @@ class sanphamController extends Controller
     }
     public function delete(Request $req){
         $row = $req->row;
+        if(session()->get('role')==1 && session()->get('islogin')==1 || session()->get('role')==2 && session()->get('islogin')==1 ){
         $model = DB::table('dienthoai')->where('id', '=', $row)->delete();
         return view('DienThoai/detail');
+        }
     }
 }
